fix: merge kernel+test into single self-contained file (repo convention)
